Extract artist insight helper to separate file
