+++ src/schema/v2/artist/insights.ts
@@ -7,59 +7,9 @@ import {
   GraphQLObjectType,
   GraphQLString,
 } from "graphql"
-import { compact } from "lodash"
 import { ResolverContext } from "types/graphql"
 import { ArtistType } from "../artist"
-
-const ARTIST_INSIGHT_KINDS = {
-  SOLO_SHOW: { value: "SOLO_SHOW" },
-  GROUP_SHOW: { value: "GROUP_SHOW" },
-  COLLECTED: { value: "COLLECTED" },
-  REVIEWED: { value: "REVIEWED" },
-  BIENNIAL: { value: "BIENNIAL" },
-  ACTIVE_SECONDARY_MARKET: { value: "ACTIVE_SECONDARY_MARKET" },
-} as const
-
-type ArtistInsightKind = keyof typeof ARTIST_INSIGHT_KINDS
-
-const ARTIST_INSIGHT_MAPPING = {
-  SOLO_SHOW: {
-    label: "Solo show at a major institution",
-    description: null,
-    fieldName: "solo_show_institutions",
-    delimiter: "|",
-  },
-  GROUP_SHOW: {
-    label: "Group show at a major institution",
-    description: null,
-    fieldName: "group_show_institutions",
-    delimiter: "|",
-  },
-  COLLECTED: {
-    label: "Collected by a major institution",
-    description: null,
-    fieldName: "collections",
-    delimiter: "\n",
-  },
-  REVIEWED: {
-    label: "Reviewed by a major art publication",
-    description: null,
-    fieldName: "review_sources",
-    delimiter: "|",
-  },
-  BIENNIAL: {
-    label: "Included in a major biennial",
-    description: null,
-    fieldName: "biennials",
-    delimiter: "|",
-  },
-  ACTIVE_SECONDARY_MARKET: {
-    label: "Active Secondary Market",
-    description: "Recent auction results in the Artsy Price Database",
-    fieldName: "active_secondary_market",
-    delimiter: null,
-  },
-} as const
+import { getArtistInsights, ARTIST_INSIGHT_KINDS } from "./helpers"
 
 export const ArtistInsightKind = new GraphQLEnumType({
   name: "ArtistInsightKind",
@@ -117,51 +67,3 @@ export const ArtistInsights: GraphQLFieldConfig<any, ResolverContext> = {
     return insights.filter((insight) => kind.includes(insight.type))
   },
 }
-
-export const getArtistInsights = (artist) =>
-  compact(
-    (Object.entries(ARTIST_INSIGHT_MAPPING) as [
-      ArtistInsightKind,
-      typeof ARTIST_INSIGHT_MAPPING[ArtistInsightKind]
-    ][]).map(([kind, { label, description, fieldName, delimiter }]) => {
-      const value = artist[fieldName]
-
-      if (!value) {
-        return { artist }
-      }
-
-      switch (typeof value) {
-        case "string":
-          const trimmed = value.trim()
-
-          if (!trimmed) return null
-
-          const entities = trimmed
-            .split(delimiter ?? "|")
-            .map((entity) => entity.trim())
-          return {
-            entities,
-            count: entities.length,
-            label,
-            type: kind,
-            kind,
-            description,
-            artist,
-          }
-
-        case "boolean":
-          return {
-            entities: [],
-            label,
-            type: kind,
-            kind,
-            description,
-            count: value ? 1 : 0,
-            artist,
-          }
-
-        default:
-          return { count: 0, artist }
-      }
-    })
-  )
